tests/wpt/web-platform-tests/referrer-policy/css-integration/image/internal-import-stylesheet.html

<!DOCTYPE html>
<html>
  <head>
    <title>CSS integration - image from imported stylesheet (internal)</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/common/utils.js"></script>
    <!-- Common global functions for referrer-policy tests. -->
    <script src="/common/security-features/resources/common.js"></script>
    <script src="/referrer-policy/generic/referrer-policy-test-case.js"></script>
    <meta name="referrer" content="origin">
  </head>
  <body>
    <p>Check that resources from stylesheets (imported from internal
    stylesheets) are loaded with the referrer and referrer policy from the
    document.</p>

    <div class="styled"></div>

    <script>
      promise_test(function(css_test) {
        var id = token();
        var url_prefix =  location.protocol + "//www1." + location.hostname + ":" + location.port + "/common/security-features/subresource/";
        var css_url = url_prefix + "stylesheet.py?id=" + id;
        var img_url = url_prefix + "image.py?report-headers&id=" + id;

        var style = document.createElement("style");
        style.type = 'text/css';
        style.appendChild(document.createTextNode("@import url('" + css_url + "');"));
        document.head.appendChild(style);
        return timeoutPromise(css_test, 1000)
          .then(() => requestViaXhr(img_url))
          .then(function(message) {
              assert_own_property(message, "headers");
              assert_own_property(message, "referrer");
              assert_equals(message.referrer, css_url);
            });
      }, "Image from imported stylesheet (internal).");
    </script>

    <div id="log"></div>
  </body>
</html>
